client/refactor: make the template selection card reusable
