+++ apps/client/src/widgets/collections/board/index.tsx
@@ -823,7 +823,6 @@ export default function BoardView({ note: parentNote, noteIds, viewConfig, saveC
                                 />
                                 <BoardProperties
                                     api={api}
-                                    available={availableTemplates}
                                     note={parentNote}
                                     shown={isEditingProperties}
                                     onClose={() => setIsEditingProperties(false)}
