--- apps/client/src/widgets/collections/board/properties.css
@@ -1,13 +1,0 @@
-.modal.board-properties-dialog {
-  /* The name takes what room there is, so the way to remove an entry stands at its trailing edge. */
-  .board-template-name {
-    flex-grow: 1;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-
-  .board-template-remove {
-    flex-shrink: 0;
-  }
-}
